// src/modules/leave/enums/leave.enum.ts
export enum ENUM_LEAVE_REQUEST_STATUS {
    PENDING = 'pending',
    APPROVED = 'approved',
    REJECTED = 'rejected',
    CANCELLED = 'cancelled',
}

export enum ENUM_LEAVE_HALF_DAY {
    FULL = 'full',
    AM = 'am',
    PM = 'pm',
}

export enum ENUM_LEAVE_ACCRUAL_METHOD {
    ANNUAL = 'annual',
    MONTHLY = 'monthly',
    NONE = 'none',
}

export enum ENUM_PROBATION_ENTITLEMENT_METHOD {
    PRO_RATA = 'pro_rata',
    FULL = 'full',
    NONE = 'none',
}